+++ ft_strlcat.c
@@ -1,8 +1,8 @@
 #include "libft.h"
 size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-	unsigned int	src_len;
-	unsigned int	dst_len;
+	size_t	src_len;
+	size_t	dst_len;
 
 	dst_len = 0;
 	src_len = ft_strlen(src);
